requirements.txt
# Core ML and Data Processing
pandas>=1.5.0
numpy>=1.21.0
scikit-learn==1.3.2
joblib>=1.2.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0

# Monitoring and Logging
structlog>=22.1.0
prometheus-client>=0.15.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Documentation
mkdocs>=1.4.0
mkdocs-material>=8.5.0
pydantic-settings>=2.0.0
